refactor: move layer/source forms to features/layers/forms
