+++ packages/base/src/features/layers/forms/layer/hillshadeLayerForm.tsx
@@ -2,13 +2,16 @@ import { IDict } from '@jupytergis/schema';
 import { UiSchema } from '@rjsf/utils';
 import React, { useMemo } from 'react';
 
+import { SchemaForm } from '@/src/formbuilder/objectform/SchemaForm';
+import {
+  processBaseSchema,
+  removeFormEntry,
+} from '@/src/formbuilder/objectform/schemaUtils';
+import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
 import { deepCopy } from '@/src/tools';
-import { SchemaForm } from '../SchemaForm';
-import { processBaseSchema, removeFormEntry } from '../schemaUtils';
-import { useSchemaFormState } from '../useSchemaFormState';
 import type { ILayerProps } from './layerform';
 
-export function WebGlLayerPropertiesForm(
+export function HillshadeLayerPropertiesForm(
   props: ILayerProps,
 ): React.ReactElement | null {
   const {
@@ -47,9 +50,6 @@ export function WebGlLayerPropertiesForm(
     const builtUiSchema: UiSchema = {};
     const dataCopy = deepCopy(formData);
 
-    removeFormEntry('color', formData, schema, builtUiSchema);
-    removeFormEntry('symbologyState', formData, schema, builtUiSchema);
-
     processBaseSchema(
       dataCopy,
       schema,
@@ -60,11 +60,12 @@ export function WebGlLayerPropertiesForm(
 
     if (schema.properties?.source) {
       const availableSources = model.getSourcesByType(sourceType);
-
       (schema.properties.source as IDict).enumNames =
         Object.values(availableSources);
       (schema.properties.source as IDict).enum = Object.keys(availableSources);
     }
+
+    builtUiSchema.shadowColor = { 'ui:widget': 'color' };
 
     return builtUiSchema;
   }, [schema, formData, formContext, model, sourceType]);
