+++ packages/base/src/features/layers/forms/layer/layerform.tsx
@@ -1,14 +1,40 @@
-import { IDict } from '@jupytergis/schema';
+/**
+ * Base (default) layer form and props.
+ * Used for RasterLayer and any layer type without a dedicated form.
+ */
+import { IDict, SourceType } from '@jupytergis/schema';
+import { Signal } from '@lumino/signaling';
 import { UiSchema } from '@rjsf/utils';
 import React, { useMemo } from 'react';
 
+import { SchemaForm } from '@/src/formbuilder/objectform/SchemaForm';
+import {
+  processBaseSchema,
+  removeFormEntry,
+} from '@/src/formbuilder/objectform/schemaUtils';
+import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
 import { deepCopy } from '@/src/tools';
-import { SchemaForm } from '../SchemaForm';
-import { processBaseSchema, removeFormEntry } from '../schemaUtils';
-import { useSchemaFormState } from '../useSchemaFormState';
-import type { ILayerProps } from './layerform';
+import type { IBaseFormProps } from '@/src/types';
 
-export function HillshadeLayerPropertiesForm(
+export interface ILayerProps extends IBaseFormProps {
+  /**
+   * The source type for the layer
+   */
+  sourceType: SourceType;
+
+  /**
+   * The signal emitted when the attached source form has changed, if it exists
+   */
+  sourceFormChangedSignal?: Signal<any, IDict<any>>;
+
+  /**
+   * Configuration options for the dialog, including settings for layer data, source data,
+   * and other form-related parameters.
+   */
+  dialogOptions?: any;
+}
+
+export function LayerPropertiesForm(
   props: ILayerProps,
 ): React.ReactElement | null {
   const {
@@ -46,7 +72,6 @@ export function HillshadeLayerPropertiesForm(
   const uiSchema = useMemo(() => {
     const builtUiSchema: UiSchema = {};
     const dataCopy = deepCopy(formData);
-
     processBaseSchema(
       dataCopy,
       schema,
@@ -61,8 +86,6 @@ export function HillshadeLayerPropertiesForm(
         Object.values(availableSources);
       (schema.properties.source as IDict).enum = Object.keys(availableSources);
     }
-
-    builtUiSchema.shadowColor = { 'ui:widget': 'color' };
 
     return builtUiSchema;
   }, [schema, formData, formContext, model, sourceType]);
