+++ packages/base/src/features/layers/forms/source/geotiffsource.tsx
@@ -3,11 +3,14 @@ import { showErrorMessage } from '@jupyterlab/apputils';
 import { RJSFSchema, UiSchema } from '@rjsf/utils';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
+import { SchemaForm } from '@/src/formbuilder/objectform/SchemaForm';
+import { FileSelectorWidget } from '@/src/formbuilder/objectform/fileselectorwidget';
+import {
+  processBaseSchema,
+  removeFormEntry,
+} from '@/src/formbuilder/objectform/schemaUtils';
+import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
 import { deepCopy, getMimeType } from '@/src/tools';
-import { SchemaForm } from '../SchemaForm';
-import { FileSelectorWidget } from '../fileselectorwidget';
-import { processBaseSchema, removeFormEntry } from '../schemaUtils';
-import { useSchemaFormState } from '../useSchemaFormState';
 import type { ISourceFormProps } from './sourceform';
 
 export function GeoTiffSourcePropertiesForm(
