+++ packages/base/src/features/layers/forms/source/sourceform.tsx
@@ -1,38 +1,40 @@
 /**
- * Base (default) layer form and props.
- * Used for RasterLayer and any layer type without a dedicated form.
+ * Base (default) source form and props.
+ * Used for RasterSource and any source type without a dedicated form.
  */
 import { IDict, SourceType } from '@jupytergis/schema';
 import { Signal } from '@lumino/signaling';
 import { UiSchema } from '@rjsf/utils';
 import React, { useMemo } from 'react';
 
+import { SchemaForm } from '@/src/formbuilder/objectform/SchemaForm';
+import {
+  processBaseSchema,
+  removeFormEntry,
+} from '@/src/formbuilder/objectform/schemaUtils';
+import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
 import { deepCopy } from '@/src/tools';
 import type { IBaseFormProps } from '@/src/types';
-import { SchemaForm } from '../SchemaForm';
-import { processBaseSchema, removeFormEntry } from '../schemaUtils';
-import { useSchemaFormState } from '../useSchemaFormState';
 
-export interface ILayerProps extends IBaseFormProps {
+export interface ISourceFormProps extends IBaseFormProps {
   /**
-   * The source type for the layer
+   * The source type for this form.
    */
   sourceType: SourceType;
 
   /**
-   * The signal emitted when the attached source form has changed, if it exists
+   * The signal emitted when the source form has changed.
    */
   sourceFormChangedSignal?: Signal<any, IDict<any>>;
 
   /**
-   * Configuration options for the dialog, including settings for layer data, source data,
-   * and other form-related parameters.
+   * Configuration options for the dialog, including settings for source data and other parameters.
    */
   dialogOptions?: any;
 }
 
-export function LayerPropertiesForm(
-  props: ILayerProps,
+export function SourcePropertiesForm(
+  props: ISourceFormProps,
 ): React.ReactElement | null {
   const {
     schema: schemaProp,
@@ -41,8 +43,8 @@ export function LayerPropertiesForm(
     model,
     filePath,
     formContext,
-    sourceType,
     dialogOptions,
+    cancel,
     formErrorSignal,
   } = props;
 
@@ -58,10 +60,10 @@ export function LayerPropertiesForm(
     schemaProp,
     model,
     syncData,
-    cancel: props.cancel,
+    cancel,
     onAfterChange: dialogOptions
       ? (data: IDict) => {
-          dialogOptions.layerData = { ...data };
+          dialogOptions.sourceData = { ...data };
         }
       : undefined,
   });
@@ -77,15 +79,8 @@ export function LayerPropertiesForm(
       removeFormEntry,
     );
 
-    if (schema.properties?.source) {
-      const availableSources = model.getSourcesByType(sourceType);
-      (schema.properties.source as IDict).enumNames =
-        Object.values(availableSources);
-      (schema.properties.source as IDict).enum = Object.keys(availableSources);
-    }
-
     return builtUiSchema;
-  }, [schema, formData, formContext, model, sourceType]);
+  }, [schema, formData, formContext]);
 
   if (!hasSchema) {
     return null;
