+++ packages/base/src/formbuilder/index.ts
@@ -1,4 +1,4 @@
 export * from './objectform/SchemaForm';
-export * from './objectform/source';
-export * from './objectform/layer';
+export * from '../features/layers/forms/source';
+export * from '../features/layers/forms/layer';
 export * from './creationform';
