+++ packages/base/src/formbuilder/objectform/components/WmsTileSourceUrlInput.tsx
@@ -1,7 +1,7 @@
 import { WidgetProps } from '@rjsf/utils';
 import React, { ChangeEvent, useState } from 'react';
 
-import { WMS_AVAILABLE_LAYERS_CACHE } from '@/src/formbuilder/objectform/source';
+import { WMS_AVAILABLE_LAYERS_CACHE } from '@/src/features/layers/forms/source';
 import { Button } from '@/src/shared/components/Button';
 import { Input } from '@/src/shared/components/Input';
 import { GlobalStateDbManager } from '@/src/shared/store';
